requirements.txt
asgiref==3.8.1
CacheControl==0.14.1
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
lxml==5.3.0
lxml_html_clean==0.4.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2==2.9.10
pygame==2.6.1
Pygments==2.19.1
pymemcache==4.0.0
pywebcopy==7.0.2
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.0
whitenoise==6.9.0
